(refactor) move all GraphQL queries & HOCs to lib folder
